fix(autosave): absorb architecture and placeholder baselines (#992)
