+++ apps/desktop/renderer/src/i18n/locales/en.json
@@ -936,7 +936,8 @@
       "manageSubscription": "Manage Subscription",
       "dangerZone": "Danger Zone",
       "deleteAccount": "Delete Account",
-      "deleteAccountDescription": "Permanently delete your account and all associated data."
+      "deleteAccountDescription": "Permanently delete your account and all associated data.",
+      "comingSoonTooltip": "Account features coming soon"
     },
     "general": {
       "zhCN": "Chinese (Simplified)",
@@ -1543,9 +1544,12 @@
       "title": "Unexpected system error",
       "description": "Some features may be affected. Please retry the operation. If the problem persists, restart the app."
     }
-  }
-,
+  },
   "versionControl": {
     "restoreComingSoon": "Version restore coming soon"
+  },
+  "common": {
+    "comingSoon": "Coming soon",
+    "featureInDevelopment": "This feature is under development"
   }
 }
